Added code for using the systick interrupt to read in adc inputs
filled in functions for initiating LED and button. Setup the adc to read
on the systick interrupts every 1ms...just for testing. Setup the rest
of the code to use timer6 for the adc and the systick timer for the
button (which will be used to turn the system on and off)